Bugfixes and SDL
- fix SDL setup to allow building on windows
- fix a bunch of things and add output functions to get IBM logo working
- add console mode for emulator

+++ include/SKChip8/Utils/ROMLoader.h
@@ -23,7 +23,7 @@ namespace SKChip8
 
     private:
         std::string filename_;
-        std::vector<uint16_t> buffer_;
+        std::vector<uint8_t> buffer_;
     };
 
 }
